requirements.txt
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==2.0.0
cryptography==45.0.7
greenlet==3.2.4
h11==0.16.0
idna==3.10
numpy==2.3.3
outcome==1.3.0.post0
pandas==2.3.2
paramiko==2.12.0
pycparser==2.23
PyMySQL==1.1.2
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
selenium==4.35.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
sshtunnel==0.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
